requirements.txt
Flask==3.0.3
pypdf==5.1.0
reportlab==4.2.5
python-dateutil==2.9.0.post0
gspread==6.1.2
google-auth==2.35.0
